requirements.txt
fastapi
uvicorn
torch
transformers
pdf2image
Pillow
huggingface_hub
pytesseract
timm
pandas
